Admin Panel
Began creating the admin panel of the site.

Functions that work:

Can add a simple post by following a script. <- Next on the list to be reworked to work dynamically on the page for a user.

Can delete posts. The only functionality currently implemented is to delete posts and then refresh the table of blog posts. Will add more functionality to edit posts that has been created.