Basic styling of the blogfeed
Began styling the blog feed page. This will be changed in future builds as I am not just working on the layout and adding responsiveness to the page so everything flows nicely.

there will be changes to the data being fetched from the api to include an image for each blog post. When that issue is resolved I will finalize the styling of the blogfeed.